+++ src/main.rs
@@ -1,6 +1,12 @@
+mod platforms;
+mod player;
+
+use platforms::PlatformsPlugin;
+use player::PlayerPlugin;
+
 use bevy::{prelude::*, sprite::MaterialMesh2dBundle, window::WindowResolution};
 use bevy_rapier2d::{
-    control::{KinematicCharacterController, KinematicCharacterControllerOutput},
+    control::KinematicCharacterController,
     dynamics::RigidBody,
     geometry::Collider,
     plugin::{NoUserData, RapierPhysicsPlugin},
@@ -9,47 +15,13 @@ use bevy_rapier2d::{
 
 const COLOR_BACKGROUND: Color = Color::rgb(0.29, 0.31, 0.41);
 const COLOR_FLOOR: Color = Color::rgb(0.45, 0.55, 0.66);
-const COLOR_PLATFORM: Color = Color::rgb(0.13, 0.13, 0.23);
 const COLOR_PLAYER: Color = Color::rgb(0.60, 0.55, 0.60);
 const FLOOR_THICKNESS: f32 = 10.0;
-const MAX_JUMP_HEIGHT: f32 = 230.0;
-const PLAYER_VELOCITY_X: f32 = 400.0;
-const PLAYER_VELOCITY_Y: f32 = 850.0;
-const WINDOW_BOTTOM_Y: f32 = WINDOW_HEIGHT / -2.0;
+
 const WINDOW_HEIGHT: f32 = 720.0;
 const WINDOW_LEFT_X: f32 = WINDOW_WIDTH / -2.0;
+pub const WINDOW_BOTTOM_Y: f32 = WINDOW_HEIGHT / -2.0;
 const WINDOW_WIDTH: f32 = 1024.0;
-
-#[derive(Component)]
-struct Jump(f32);
-
-#[derive(Bundle)]
-struct PlatformBundle {
-    sprite_bundle: SpriteBundle,
-    body: RigidBody,
-    collider: Collider,
-}
-
-impl PlatformBundle {
-    fn new(x: f32, scale: Vec3) -> Self {
-        Self {
-            sprite_bundle: SpriteBundle {
-                sprite: Sprite {
-                    color: COLOR_PLATFORM,
-                    ..Default::default()
-                },
-                transform: Transform {
-                    translation: Vec3::new(x, WINDOW_BOTTOM_Y + (scale.y / 2.0), 0.0),
-                    scale,
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-            body: RigidBody::Fixed,
-            collider: Collider::cuboid(0.5, 0.5),
-        }
-    }
-}
 
 fn main() {
     App::new()
@@ -65,96 +37,10 @@ fn main() {
         }))
         .add_plugin(RapierPhysicsPlugin::<NoUserData>::pixels_per_meter(200.0)) // Physics plugin
         .add_plugin(RapierDebugRenderPlugin::default()) // Debug pl
+        .add_plugin(PlatformsPlugin)
+        .add_plugin(PlayerPlugin)
         .add_startup_system(setup)
-        .add_system(movement) // new system added
-        .add_system(jump) // new system added
-        .add_system(rise) // new system added
-        .add_system(fall) // new system added
         .run();
-}
-
-fn fall(time: Res<Time>, mut query: Query<&mut KinematicCharacterController, Without<Jump>>) {
-    if query.is_empty() {
-        return;
-    }
-
-    let mut player = query.single_mut();
-
-    // I am using two-thirds of the Y-velocity since I want the character to fall slower than it rises
-    let movement = time.delta().as_secs_f32() * (PLAYER_VELOCITY_Y / 1.5) * -1.0;
-
-    match player.translation {
-        Some(vec) => player.translation = Some(Vec2::new(vec.x, movement)),
-        None => player.translation = Some(Vec2::new(0.0, movement)),
-    }
-}
-
-fn jump(
-    input: Res<Input<KeyCode>>,
-    mut commands: Commands,
-    query: Query<
-        (Entity, &KinematicCharacterControllerOutput),
-        (With<KinematicCharacterController>, Without<Jump>),
-    >,
-) {
-    if query.is_empty() {
-        return;
-    }
-
-    let (player, output) = query.single();
-
-    if input.pressed(KeyCode::Up) && output.grounded {
-        commands.entity(player).insert(Jump(0.0));
-    }
-}
-
-fn rise(
-    mut commands: Commands,
-    time: Res<Time>,
-    mut query: Query<(Entity, &mut KinematicCharacterController, &mut Jump)>,
-) {
-    if query.is_empty() {
-        return;
-    }
-
-    let (entity, mut player, mut jump) = query.single_mut();
-
-    let mut movement = time.delta().as_secs_f32() * PLAYER_VELOCITY_Y;
-
-    if movement + jump.0 >= MAX_JUMP_HEIGHT {
-        movement = MAX_JUMP_HEIGHT - jump.0;
-        commands.entity(entity).remove::<Jump>();
-    }
-
-    jump.0 += movement;
-
-    match player.translation {
-        Some(vec) => player.translation = Some(Vec2::new(vec.x, movement)),
-        None => player.translation = Some(Vec2::new(0.0, movement)),
-    }
-}
-
-fn movement(
-    input: Res<Input<KeyCode>>,
-    time: Res<Time>,
-    mut query: Query<&mut KinematicCharacterController>,
-) {
-    let mut player = query.single_mut();
-
-    let mut movement = 0.0;
-
-    if input.pressed(KeyCode::Right) {
-        movement += time.delta_seconds() * PLAYER_VELOCITY_X;
-    }
-
-    if input.pressed(KeyCode::Left) {
-        movement += time.delta_seconds() * PLAYER_VELOCITY_X * -1.0;
-    }
-
-    match player.translation {
-        Some(vec) => player.translation = Some(Vec2::new(movement, vec.y)), // update if it already exists
-        None => player.translation = Some(Vec2::new(movement, 0.0)),
-    }
 }
 
 fn setup(
@@ -162,10 +48,6 @@ fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
-    commands.spawn(PlatformBundle::new(-100.0, Vec3::new(75.0, 200.0, 1.0)));
-    commands.spawn(PlatformBundle::new(100.0, Vec3::new(50.0, 350.0, 1.0)));
-    commands.spawn(PlatformBundle::new(350.0, Vec3::new(150.0, 250.0, 1.0)));
-
     commands.spawn(Camera2dBundle::default());
 
     commands
