+++ src/player.rs
@@ -1,0 +1,108 @@
+use bevy::prelude::*;
+use bevy_rapier2d::prelude::*;
+
+const MAX_JUMP_HEIGHT: f32 = 230.0;
+const PLAYER_VELOCITY_X: f32 = 400.0;
+const PLAYER_VELOCITY_Y: f32 = 850.0;
+
+#[derive(Component)]
+
+struct Jump(f32);
+
+pub struct PlayerPlugin;
+
+impl Plugin for PlayerPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_startup_system(setup)
+            .add_system(movement)
+            .add_system(jump)
+            .add_system(rise)
+            .add_system(fall);
+    }
+}
+
+fn setup(_commands: Commands) {}
+
+fn fall(time: Res<Time>, mut query: Query<&mut KinematicCharacterController, Without<Jump>>) {
+    if query.is_empty() {
+        return;
+    }
+
+    let mut player = query.single_mut();
+
+    // I am using two-thirds of the Y-velocity since I want the character to fall slower than it rises
+    let movement = time.delta().as_secs_f32() * (PLAYER_VELOCITY_Y / 1.5) * -1.0;
+
+    match player.translation {
+        Some(vec) => player.translation = Some(Vec2::new(vec.x, movement)),
+        None => player.translation = Some(Vec2::new(0.0, movement)),
+    }
+}
+
+fn jump(
+    input: Res<Input<KeyCode>>,
+    mut commands: Commands,
+    query: Query<
+        (Entity, &KinematicCharacterControllerOutput),
+        (With<KinematicCharacterController>, Without<Jump>),
+    >,
+) {
+    if query.is_empty() {
+        return;
+    }
+
+    let (player, output) = query.single();
+
+    if input.pressed(KeyCode::Up) && output.grounded {
+        commands.entity(player).insert(Jump(0.0));
+    }
+}
+
+fn rise(
+    mut commands: Commands,
+    time: Res<Time>,
+    mut query: Query<(Entity, &mut KinematicCharacterController, &mut Jump)>,
+) {
+    if query.is_empty() {
+        return;
+    }
+
+    let (entity, mut player, mut jump) = query.single_mut();
+
+    let mut movement = time.delta().as_secs_f32() * PLAYER_VELOCITY_Y;
+
+    if movement + jump.0 >= MAX_JUMP_HEIGHT {
+        movement = MAX_JUMP_HEIGHT - jump.0;
+        commands.entity(entity).remove::<Jump>();
+    }
+
+    jump.0 += movement;
+
+    match player.translation {
+        Some(vec) => player.translation = Some(Vec2::new(vec.x, movement)),
+        None => player.translation = Some(Vec2::new(0.0, movement)),
+    }
+}
+
+fn movement(
+    input: Res<Input<KeyCode>>,
+    time: Res<Time>,
+    mut query: Query<&mut KinematicCharacterController>,
+) {
+    let mut player = query.single_mut();
+
+    let mut movement = 0.0;
+
+    if input.pressed(KeyCode::Right) {
+        movement += time.delta_seconds() * PLAYER_VELOCITY_X;
+    }
+
+    if input.pressed(KeyCode::Left) {
+        movement += time.delta_seconds() * PLAYER_VELOCITY_X * -1.0;
+    }
+
+    match player.translation {
+        Some(vec) => player.translation = Some(Vec2::new(movement, vec.y)), // update if it already exists
+        None => player.translation = Some(Vec2::new(movement, 0.0)),
+    }
+}
